removed "Hello, " on small screen
added class "large_screen"
The concept is that any time you want to target something for a
particular device - use a class, such as "large_screen" or "mobile" and
in css use @media to turn on or off by display: hidden.

Also, I want to use generators as generators, rather than "outputters",
so they can be combined in flexible ways. _span is a start. The drawback
is a lot more 'echo' in code