Fix incorrect base_url
On pages bli-medlem and aktivering the base url is incorrect where it points to e.g. http://localhost/aktivering/ instead of http://localhost/. This makes the language image not to show correctly.
This is also a security fix, HTTP_HOST variable can be modified by the user request, so best practice is to avoid using it or at least have a white list if used.
More info about security issues can be found on http://shiflett.org/blog/2006/mar/server-name-versus-http-host